tests/wpt/tests/css/css-lists/li-counter-increment-computed-style.html

<!doctype html>
<title>Magic list-item counter-increment shouldn't be visible from computed style</title>
<link rel="help" href="https://drafts.csswg.org/css-lists/#declaring-a-list-item">
<link rel="help" href="https://drafts.csswg.org/css-lists/#list-item-counter">
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<li data-expected="none">No explicit counter.
<li><span style="counter-increment:inherit" data-expected="none">Inherited
<li value="10" data-expected="none">Value attribute.
<li style="counter-increment: list-item 10" data-expected="list-item 10">Explicit list-item counter.
<li style="counter-increment: list-item 1" data-expected="list-item 1">Explicit and redundant list-item counter.
<li style="counter-increment: foo 10" data-expected="foo 10">Other counter.
<script>
test(function() {
  for (const element of document.querySelectorAll("[data-expected]"))
    assert_equals(getComputedStyle(element).counterIncrement, element.getAttribute("data-expected"), element.innerText);
}, "list-item counter-increment shouldn't be visible from computed style");
</script>
