LayoutTests/imported/w3c/web-platform-tests/css/css-lists/li-counter-increment-computed-style-expected.txt

No explicit counter.
Inherited
Value attribute.
Explicit list-item counter.
Explicit and redundant list-item counter.
Other counter.

FAIL list-item counter-increment shouldn't be visible from computed style assert_equals: Inherited expected "none" but got ""
